requirements.txt
streamlit==1.24.1
pandas>2.0
numpy<=1.25.1
streamlit-flow-component
streamlit-d3graph
